+++ CHANGES.rst
@@ -11,6 +11,12 @@ Changelog of fews-3di
 
 - Added evaporation upload.
 
+- Simulation is actually being run now.
+
+- Added processing of the results.
+
+- Added usage instructions.
+
 
 0.1 (2020-04-09)
 ----------------